Rename groupMembers.js to groupTasks.js
