+++ src/views/Profile/index.js
@@ -15,7 +15,7 @@ import {
 import ListItem from 'components/ListItem'
 import { ITEM_TYPES, COMPLETION_STATUS, AGE_GROUPS } from 'consts'
 import CompletedTasks from './CompletedTasks'
-import { getTaskGroupsWithChildTaskGroups } from '../../helpers/groupMembers'
+import { getTaskGroupsWithChildTaskGroups } from '../../helpers/groupTasks'
 
 const Background = styled.div`
   min-height: 100vh;
